Made panels file fully PEP8 compliant.
I also commented out some panels that weren't being used. Previously, it
didn't matter, because they weren't being initialized in __init__.py,
but now they are since we switched to "per module registration." For a
similar reason, I had to move the Model Objects panel in the code so
that it would appear where it was orignally.