Hid advanced molecule options.
Target only, custom time step, and custom space step are advanced
options that basic users may not need, so they have been hidden under a
sub-panel. This isn't a a true panel, but it is collapseable just like a
real one. It might be useful to add something similar to this for other
"advanced" options to prevent new users from becoming overwhelmed or
confused.