Molecules are now created, validated, then added.
Previously, we added a molecule, then modified it to make it valid. Now,
the molecule doesn't actually exist as part of an MCell model until it
every part of it has been validated and explicity added. This is still
fairly broken. The molecules add correctly (for the most part), but
there's no way to select/load them after they've been created.