Molecules can be added, removed, or updated.
The details of how the UI is organized still needs ironed out, but this
seems like a big improvement over the previous version. Everything is
mostly functional, except that the code is not copying the advanced
molecule parameters (e.g. target only).

Error checking only happens at the time of adding or updating now.
This is just a test to see if we like it better this way. The update
also gives positive feedback to let the user know that molecules have
been added, removed, or updated.