Slightly rearranged UI in define molecule panel.
Moved template_list to the beginning of the panel. Otherwise, it always
jumps around as various conditional fields appear and disappear. The
"Delete Molecule" button (previously "Remove from List") was grouped
with the other buttons. As such, it only appears when you select
something in the list, which seems preferable.