Partitions can be defined and exported now.
The range and step length of the X, Y, and Z axes can be set in a new
Partitions panel. The user can also display a visual representation of
the partition boundaries. A wireframe view of a cube is used for this
purpose. Originally, I thought a lattice would be a natural fit for
this, as lattices and partitions both consist of subvolumes. However,
there is a relatively small limit on the number of subvolumes that can
exist in a lattice. As such, it seems better to only display the
partition boundaries and to not bother showing the subvolumes at all.