Added comments. Fixed export and viz bugs.
Comments and docstrings were added to the operators file.

The visualization bug occurred when viz data was first imported and the
cursor was not at the origin. The offset of the origin was applied to
both the positions and the shape/glyph. Offsetting the positions could
be seen as a case of "not a bug, but a feature." However, this also
caused the positions and shapes to be offset from each other, which
clearly seems like a bug. Also, I think it's fairly standard practice to
have importers center objects at the origin.

The export bug has to do with a matrix multiplication problem. I believe
that starting with Blender 2.62, the order of matrix multiplication was
reversed. This bug caused some meshes to be exported in the wrong
orientation.

A number of other small changes were made:
- Added a link to the google code tracker
- Icons added for prop_search elements
- Renamed variables for consistency and clarity
- More PEP8 changes
- Commented out some unused code