Deselect list item after operation (e.g. add).
Previously, the item appeared to be selected in the list, even though
the options for deleting or updating that molecule weren't available
until you clicked on the item in the list. Now, when you add, update,
cancel, or delete a molecule, nothing is selected in the list after the
operation is complete.

Also, I added a box around the advanced options to be consistent with
the changes in the main branch.