Added a changelog.
This changelog doesn't exactly follow the standard formatting, but it
seemed like it would be more useful to users this way. I broke it down
according to "features added" and "bugs fixed" since the last release on
the website. It might also be nice to have a section outlining the more
technical changes that were made for the sake of the developers (e.g.
the switch to per module registration).