Updated installation instructions in README.
The new instructions are the same across all platforms and don't require
manually copying the addon into a (potentially hidden) directory.
Rather, it is all done through Blender's UI.

The rest of the README was also updated to reflect recent changes in
CellBlender. The usage section is still not comprehensive, but the
tutorials on the website should probably suffice for the time-being.