Fixed several bugs in exporter. Updated init info.
Previously, if you exported a CellBlender project when there was nothing
in the "Model Objects" list, an include statement would still be created
for it, which would generate an error when the main MDL was run. Also,
the surface class and modify surface region files were not being
exported when the user specified "Single Unified MCell MDL File" as the
"Export Format". These bugs have been fixed, and the comments in this
file were formatted for consistency.

In __init__.py, the Blender version and API were updated.