Temporary fix for molecule directionality problem
This is a temporary fix to the issue listed here:

https://code.google.com/p/cellblender/issues/detail?id=10

To recap, foreach_set is broken for normals in 2.66. In practical terms
for CellBlender, this means that the directionality of molecule
glyphs/shapes is not being set correctly.

We should revert back to the original method once the bug is fixed in an
official Blender release, since the workaround is slower. Hopefully,
the fix will appear in Blender 2.66a.