Revert to previous molecule directionality method
Blender 2.66a has been released with a fix for the foreach_set bug.
Therefore, the workaround is no longer needed, so we can revert back to
the original method of applying molecule directionality. The version and
revision number were also updated. Bob pointed out that the version
number was being formatted incorrectly. This has been corrected.