MCell simulations can run inside CellBlender now.
Instead of having to run MCell from the command line, this allows it to
be run inside of CellBlender. First, the user has to navigate to and set
the MCell binary to use. This can be saved in the startup file, so it
doesn't need to be specified every time Blender starts. The range of
seed values can also be set. Additional parameters could be added like
the ability to log the output to a file or disable it altogether.
Currently, there is a problem where MCell will prevent the user from
interacting with Blender until it has finished running.