Multiprocessing support added to sim runner.
The simulation runner can now run multiple MCell simulations with
different seed values at the same time. This implemenation uses both
multiprocessing and threading. Given that I'm somewhat inexperienced
with both of these libraries, there may very well be a better way to do
this. It seems sufficient for basic tasks though.

The number of simultaneous MCell process can be specified. The default
and maximum value is set to the number of cores on the current machine.

The "Export and Run" panel was split up into the "Export Project" and
"Run Simulation" Panels. Several other small things in these sections
were shifted around for the sake of clarity and visibility.

The simulation runner needs some error handing and a way to cancel
currently running MCell simulations.