Added a property for setting the path to the MCell executable.
I created a string property named "MCell Path" and added a control
to set it in the MCell Project Panel. The control only seems to
allow choosing a directory name and not the executable itself, so
the string "/mcell" is appended to the the directory name. Note
that this assumes a UNIX path separator and should be changed for
compliance with other operating systems.