Implemented crude ability to run from project directory other than cwd
The previous version would run mcell from the current working directory
(cwd). This meant that to get proper behavior, Blender needed to be
started in the project directory so the creation of the viz_data and
react_data directories would place them in the project.

This version uses the property set by the Set CellBlender Project Directory button to
define the directory where the viz_data and react_data subdirectories
(and files) are created.

There are still some other directory related issues to be worked out.
For example, should we require the setting of the Molecule Viz
Directory, or should that default to the project directory with viz_data
appended? I know in my own simulations, I've forgotten to change the
Molecule Viz Directory when switching projects and been confused by why
the output wasn't changing. But that could easily be considered operator
error on my part. For ease of use, I'm inclined to want to tie all of
the files to a single project directory, but that limits flexibility.
Maybe the best solution would be a user interface check box that says
something like USE CUSTOM VIZ DIRECTORY. It would normally be unchecked,
but when it was checked, it would add the controls to specify a
different directory.