Removed threading, retained multiprocessing.
This simplifies things somewhat. I was using the map method before which
blocks the user from interacting with the GUI, which is why I had to use
threading originally. However, map_async works just as well for our
purposes.

We should probably not be using threading in Blender like this anyways
because of this:
http://www.blender.org/documentation/blender_python_api_2_62_2/info_gotcha.html#strange-errors-using-threading-module

Also, there is a problem that needs to be resolved when using
multiprocessing on Windows which is detailed here:
http://docs.python.org/2/library/multiprocessing.html#windows