Expanded Visualization Output Settings panel.
Both the range of iterations and individual molecule species can now be
set under the Visualization Output Settings panel. As an example, let's
say I have molecules A, B, and C in a simulation that lasts 1000
iterations; I could set molecules A and B (thus excluding C) to be
exported over the 200th to 500th iterations.

Fixed a bug where reaction and viz data output were not exported when
the Export Format was set to Single Unified MCell MDL File.

Changed formatting of reaction and viz data output mdls to be consistent
with the other mdls (i.e. two space indent instead of four and different
curly brace placement).