Enabled running mcell in current project directory
This version uses the cwd=dir parameter of subprocess.call() to set
the location where mcell will be started. This approach allows the
MDL files to use relative paths (as they have in the past) and still
allow the opening of different project directories within Blender.
Note that this approach calls os.path.dirname(mdl_filepath) and then
uses that path as the location where mcell will be run.

This fixed the problem of generating the viz and rxn files wherever
Blender was started regardless of the project file that was opened.