Some work on SHA1 version ID of source files.
This was initially supposed to be a relatively quick change to the
master branch, but it has become a little more complicated because
the SHA1 computed during registration may be difficult to pass to the
panels where it needs to be displayed. So while this version is mostly
working, it isn't finished. So this branch should NOT be merged until it
is finished.