Crude but functioning version system based on SHA1
This is just a slightly cleaned up version of the previous commit which
implemented a SHA1 "ID" of the source code. Initially, I hadn't planned
to push it into master until it was cleaned up a bit, but I think the
need to identify versions is probably more important right now than how
gracefully it gets done. So I took the liberty of incrementing our
version number up to 0.1.56 so we'll know that all versions higher than
0.1.55 will have some form of source code identification in them.

For this version, the souce code ID is not displayed in the panels (see
previous commit message), but it is printed to the console.

Sorry it's so ugly right now...