Changed bpy.utils.script_paths()[0] back to [1]
It appears that the index of the CellBlender script path differs
on different systems. The index of [1] worked on machine "rufus"
but the index of [0] was needed on Markus' machine. This is indeed
very brittle behavior, and I'll look into fixing it. I've changed
it back to [1] for now so it will function on my machine.