Use bpy.utils.user_resource('SCRIPTS')
Use bpy.utils.user_resource('SCRIPTS') instead of
bpy.utils.script_paths()[n] to find the source files to compute the SHA1.

It appears that script_paths() doesn't return values in a predictable
way. On some installations, the 0th index contains the user installed
addons path, and on other installations the 1st index contains the user
installed addons path. This was determined by trial and error since the
documentation wasn't specific.

The documentation for bpy.utils.user_resource('SCRIPTS') is also not
specific, but since it returns a single string (rather than a list of
strings) it has a better chance of being repeatable.