Separated identify_source_version from register()
Created a new function in __init__.py for doing the identification. By
separating this functionality, it *could* be possible to call the
identification code when running standalone (__name__ == '__main__').
However, because of various other dependencies, __init__.py won't
currently run from a python command line outside of Blender. We might be
able to overcome that with "try" blocks, but it's not that important
right now, and I don't want to make that many changes to __init__.py
without some discussion.