Tolerably working ID reporting using SHA1 to a file
This version computes the SHA1 of all the source files listed in the
bl_info dictionary with the "cellblender_source_list" key. It writes
that SHA1 to a file in the cellblender addon folder with the file name
of "cellblender_source_sha1.txt". This file is read by the compute_id
function in the "MCELL_PT_project_settings" class which is called by the
draw function of that same class to display the current CellBlender ID.

There should be a better way to get the ID (which is computed inside of
__init__.py) to the panel (cellblender_panels.py), but I didn't find an
obvious way, so I had __init__ write the ID to a file to be read by the
panel (yes, every time it's redrawn). While re-reading this file every
time isn't the best way to share the ID, the file itself can be helpful
because it can be read without even running Blender. This might be
helpful for us in diagnosing user configuration problems. In particular,
if the date of that file ("cellblender_source_sha1.txt") doesn't change
when the user loads CellBlender, then that indicates that they may not
be installing CellBlender in the proper location and Blender is possibly
getting its CellBlender from a different path. If the file date does
change when CellBlender is loaded (by starting Blender, for example),
then it should be reflecting the proper SHA1 in the CellBlender Project
Settings panel.

The SHA1 reported for this version is:
  9044287420de08ece3e20e1d193f367a1ebc359e