Implemented a simple command-line plotting capability
Added a string property for a plotting command. The command is split
using python's str.split() function, and the resulting list is passed to
the subprocess command with cwd set to the project directory with
"react_data" appended.

Using an xmgrace example, the command might be:

  /usr/bin/xmgrace

Using a java program example, the command might be:

  /usr/bin/java -jar PlotData.jar fxy=v1.dat fxy=v2.dat fxy=s1.dat

The user may interact with whatever program is launched to open various
data files and change the plot settings. In the java example, file names
themselves are provided on the command line.

This is obviously a general-purpose "catch all" approach, but we should
also provide plotting controls tailored to specific situations and
plotting packages.