Bugfix: Checking for script in wrong directory
This fixes a bug wherein we only checked for CellBlender in the USER
script directory but neglected the LOCAL and SYSTEM script directories.
Furthermore, it checks for the script directories in the order that
Blender does in case the user has multiple versions of CellBlender
installed in different locations (e.g. one in USER and one in LOCAL).

Tom has suggested a better fix to Bob, but this is a quick fix in the
meantime.

I also made some PEP8 fixes.

CellBlender Source SHA1 =  0181575f6ec3a0b9c9ff72aa2f64fbd6dddcd3b9