Obtained source ID from bl_info by importing cellblender
This was a much better way to get the SHA1 computed by __init__.py into
the panels. The code still writes to the "cellblender_source_sha1.txt"
file, but this file does not need to be read (or re-read) to get the ID
into the panels on every redraw.

By the way, this brings up the question of whether the fields we're
adding to bl_info really belong in bl_info. Rather than polluting that
structure with our own fields, maybe we should add a "cellblender_info"
dictionary at that same level for our own "global" data.