Simple code to look for plotting packages and modules
This modification checks the current path for a variety of plotting
programs (python, xmgrace, java, and excel) and a few python plotting
modules (matplotlib, matplotlib.pyplot, pylab, numpy, scipy). It doesn't
do anything with any of them other than to print whether they are found
or not. All of the testing code is included in a "try/except" block to
hopefully keep it from crashing.

Eventually, if we are going to support a variety of plotting options, we
might want to create a special folder in the addon that contains all of
the installed options (python code that knows how to interface to each
plotting package). Then our CellBlender code can just scan that folder
to populate the plotting panel with installed options.

CellBlender Source ID (SHA1) =  ccb970cf34bdadbc8c736f08573bf9b82c2617b1