Added 2 RXN plotting plug-ins with hard-coded buttons
This version contains 2 buttons for plotting reactions:
  MatPlotLib ( All Reactions )
  Java Plot ( All Reactions )
These are in addition to the Custom Plot Command.

These two buttons currently expect to find the files:
  plot_simple.py
  PlotData.jar
in the "addons/cellblender/data_plotters" folder. If the
files are not found when the buttons are pressed, Blender
generates error messages but continues to run. For this
reason, I decided to commit and push these changes even
though we haven't yet decided which (if any) of these
plotting plug-ins will be considered part of CellBlender
itself. As the buttons suggest, both of them will attempt
to plot all of the reactions produced by MCell. However,
they currently assume a seed value of 0001 which appears
in the data file names (vol1.World.0001.dat).

This version also contains the new cellblender_info
dictionary to hold the values that had been previously
stuffed into the bl_info dictionary.

CBSID =  0b7b29108533f51eb552144960972794f804cadd