Some progress on isolating the plotting functionality
The current structure is contained in a directory titled "data_plotters"
and it contains an __init__.py file which searches for subdirectories
under that folder which similarly contain __init__.py files. Any that
are found are imported and their "requirements_met()" function is
called. The requirements_met function may make whatever tests are needed
to determine if the system contains the needed functionality to support
that particular plotter. It returns True if the requirements are met,
and false if they are not. Each such plotting package will contain a
small number of standardized functions ("duck typing?") to support
plotting. These standardized functions are still being worked out.

This particular commit does not yet perform the plotting any differently
than the previous version, but the infrastructure is being built to
support the more general concept of plotter plug-ins ("duck typing?") to
support plotting. These standardized functions are still being worked
out.

This particular commit does not yet perform the plotting any differently
than the previous version, but the infrastructure is being built to
support the more general concept of plotter plug-ins. This code was
written last night but not checked in until this morning.

CellBlender Source ID (SHA1) =  edbb386f999f303fbda5a319bf8850893264bac3