Non-functioning attempt to get plotters list into panels
This version tries to create a dynamic list of plotters from the names
stored in cellblender_info['cellblender_plotting_modules'], but it did
not work. Part of the problem was in creating or updating the
EnumProperty.

This is being committed to save its state, but it's likely that a
different approach will be tried next.

CellBlender Source ID (SHA1) =  1055afa16e56d7681f11b2b06fa3ae2a030219d9