Attached button labels to buttons for use by a common operator
This is the beginning of a consolidation of the plotting user interface
elements. Prior versions (and this one for now) have used a separate
operator for each plotting module. This version begins to explore the
possibility of using a single operator that works for all plotting
modules. This will make it much easier to add new plotting packages.

CellBlender Source ID (SHA1) =  f998b8802298cc9684f2484480187a2ddda5d270