Working version of plotting plug-in system
These changes support "pluggable" plotting modules. A new module is
added by creating a new directory in the "data_plotters" directory. The
new directory must contain an "__init__.py" file with the following
functions:

  get_name() - Returns a user-friendly name for this plotter

  requirements_met() - Returns True/False based on runnable status

  plot ( data_path, plot_spec ) - Requests a plotting of the data files
    in the "data_path" directory according to the plot specification in
    "plot_spec". The plot_spec is currently a string with parameters:
        keyword  or  keyword=value
    Each plotting plug-in must translate these plotting parameters into
    commands which are understood by their plotting package. The most
    basic of these keywords are:
        page  - intended to cause a "page break" between plots
        plot  - intended to create a new plot within the current page
        f=fn  - specifies a file to be plotted in current page/plot
    The simplest translation of these keywords for plotting ignores all
    page and plot commands and simply adds each new file name (fn) to
    the same plot. This is exactly what mpl_simple.py currently does.

For example implementations of this interface, see these files:

    data_plotters/xmgrace/__init__.py
    data_plotters/java_plot/__init__.py
    data_plotters/mpl_simple/__init__.py
    data_plotters/mpl_simple/mpl_simple.py
    data_plotters/mpl_plot/__init__.py
    data_plotters/mpl_plot/mpl_plot.py

Finally, note that this interface is still under development, so it is
reasonable to expect significant changes and clean-up in the near
future. Also note that many of these files are not part of the
CellBlender Source ID. At some point we should  decide which files are
part of CellBlender itself and which are considered optional plug-ins.

CellBlender Source ID (SHA1) =  69a2c796706008e2cd713737651562e6f4aef615