Added a "Combine Seeds" plotting capability
Added a checkbox and the logic needed to plot files with identical
names, but different seed values on the same plot. This is discussed
in a little more detail (with a few examples) on the wiki for issue 17.

CellBlender Source ID = 4a3a97bac31cbe4086bb359a3047d6e705049b6a