Partial implementation of relative file paths
This version contains a partial implementation of relative paths. More
specifically, all of the files should be created in the proper places,
but the default visualization directory does not currently follow the
location of the blender file and must be set manually (to be completed
by the end of tomorrow - hopefully). There may be other limitations as
well. Also, this version does not yet build the newer deeper directory
structure: Project_files/mcell/SCENE_files/SCENE.main.mdl that was
discussed today.

The test for this change is to create a .blend file in some directory,
and then save it to a new directory and export/run the model. All of the
directory paths (except the path to mcell itself) should now be relative
to the new location of the .blend file.

Also in this version, I've created a "Scratch" panel for testing various
user interface capabilities. In this case, I wanted a way to show all
icons which are currently available, so I added that capability to the
Scratch panel. I also included the ability to print the icon names in
case we'd like to search them for particular strings. Of course, this
scratch panel should be removed before major releases.

None of these changes should keep this version from functioning so it
was updated in the master branch.

CellBlender Source ID (SHA1) =  dd8749242d06e9e1450be2bda624fb92819f8162