Outputs Viz Data relative to .blend file, export needs work.
This version saves the Viz Data relative to the .blend file location,
and it exports the MDL relative to the .blend location. However, the MDL
export still brings up the file selection window and I have not been
able to have it not do that. Also, if the file selection is used to
navigate to another location, the MDL is saved there (rather than the
proper location relative to the .blend file. I inserted a lot of debug
statments to figure out how to stop the file selection panel from coming
up, but was not able to figure it out. Other than having to push one
more button than should be necessary, everything seems to be working.

I also consolidated the "Export" and "Read Viz" buttons next to the "Run
Simulation" button since that's part of the current work flow ... modify,
export, run, read viz. Some of these steps will go away as we implement
automatic exporting and automatic viz loading.

I'm sure I've violated PEP8 a few times, but this is still a work in
progress, and I can pretty it up after it works. Sorry for the mess.

CellBlender Source ID (SHA1) =  c26e88ba8b9c8166d989ad78b3dccba4c9ad33a4