Prevent file selector when exporting project.
This commit prevents the file selector from appearing when exporting the
project, thereby ensuring that it is created in the proper location
relative to the blend. I also added in a short status message, so the
user knows that the operation was successful.

CellBlender Source ID = 4f5278225639591bbaee2caaba083bf59f8d7a41