Simulation runner now runs as a separate process.
This is a much safer way of running the simulation runner. The
multiprocessing package technically requires that the __main__ module be
importable by the children. In practice, the old way worked okay on *nix
systems aside from memory problems. However, on Windows systems, it's
unuseable, as it essentially created something like a fork bomb.
Additional changes still need to be made to be fully runnable on
Windows, but this is a good first step.

CellBlender Source ID = 49996589456bae9e5bfdabdf8d49a9bee600530e