Tracks progress of simulations in template_list.
The progress of MCell simulations is displayed in a template list in the
Run Simulations panel. The functionally is somewhat limited due to the
fact that I recently had to change the simulation runner so that it runs
as a separate process, which in turn makes it trickier to track and kill
individual MCell simulations. For now, we are currently only showing the
status of a *set* of MCell simulations being run (e.g. if you are
running seeds 1-10, it is shown as one item in the list instead of 10
distinct entries). The status is shown as one of three possible states:
running, completed, or failed. Additional information shown in an entry
includes the PID, the name of the main mdl, and the seeds. Completed and
failed entries can be cleared from the list via a button in the panel.