Added initial CellBlender Preferences panel.
This adds an initial CellBlender Preferences panel. There is only one
option for now, which allows the user to enable or disable the filtering
of invalid entries. The default behavior is enabled as this is what most
users will want. However, if CellBlender ever marks something as invalid
which is actually valid in MCell the user can override it here.

There are at least two possible ways of handling CellBlender
Preferences: storing them in the startup blend file or storing them in a
unique non-blend file. After discussions with the rest of the group, the
consensus seemed to be tentatively in favor of storing settings in the
startup blend. There are advantages to having a unique non-blend
CellBlender Preferences file, but storing them in the startup blend file
is a quick and easy solution for the time being.

CellBlender Source ID = fb5af334114d1a77adc3d0044987dc674124451a