This version SEEMS to be working with the new directory structure.
This version exports to the new directory structure:

   project.blend
   project_files/
      react_data/
      viz_data/
      SceneName.main.mdl
      SceneName.geometry.mdl
      :

It also runs from the new directory structure.
It also (usually) displays viz from the new directory structure.

But for some reason, the plotting is now broken. The main reason
for pushing this change is to get the new directory structure into
our build before the code freeze (or at least the feature freeze)
planned this week.

There's a lot of clean up that I still need to do.