Fixed plotting, but found odd problem with multiple seeds.
Experiments indicate that dropping a blend file into an empty directory
works to create a runnable project. However, on the first run of that
project, the multiple runs for multiple seeds are not completed
properly. But on subsequent runs, they are. This created what appeared
to be plotting errors, but the problem was that the files had no data.
Actually, both XmGrace and the Java program handled this case and
plotted the data that they did find. But the two python versions failed
because they are relying on having data (which might also need to be
fixed).

CellBlender Source ID (SHA1) = caa9ca872f9cbbc6286a1edc3d7650f615d1f22b