Check for illegal names in Model Objects list.
Entries in the Model Objects list are now checked in much the same way
that molecules and reactions are. In other words, a legal name will
receive a green check mark in the Model Object list, whereas an illegal
one will be marked with an error icon (and will also be filtered out
from export by default).