Added error status for un-triangulated meshes.
If an object with quads is added to the Model Objects list, then it is
marked with an error status.

This requires iterating over the faces of the mesh and checking for
quads. Therefore, this makes adding objects to the Model Objects list
slightly slower for really large meshes. I personally tested it with an
object having nearly three million faces, and it took about a second.
As such, I don't think it's a problem, but we could make this a
CellBlender Preference if it becomes an issue for anyone.

CellBlender Source ID = dc8932930b3618613ff5ba45d429f5e4e3241ba7