Cleaned up function "project_files_path()" to enable sharing
This function (for expediency) was created in both the operators
and panels files. One of the two  originally needed a "context"
parameter to get the scene name, but that functionality was removed.
Since the context parameter was removed they should be identical.
So they should be merged and shared between the two files, but at
this time I'm not sure of the best place for it and the best
mechanism for sharing (my inexperience with Python is showing).

CellBlender Source ID = 1b8b7c969250b4903b46c943df8055bbb22f15bd