Cleaned up interface. Made more consistent.
Formatted how "Set Path to MCell binary" looks, so that it is more
obvious and consistent with how the Project Directory is done.

The functionality of the base project name was replaced by the scene
name, so I am displaying the scene name where the project name used to
be.

I made it so that the blend needs to be saved before the project can be
exported in the Export Project panel.

Commented out the scratch/testing panel to prepare for an official
release. Also the functionality of displaying icons is available in
another built-in Blender addon.

Made export project icons consistent.

Made some more PEP8 changes.