Checks for executable mcell
This version checks to see if the path to mcell contains an executable
file using the os.access(,os.X_OK) function and related logic.

This is checked both when a new mcell path is set and whenever the
panel is redrawn. The panel redraw check makes the mcell path setting
check somewhat redundant, but both were left in the code because the
panel redrawing might NOT be the best place to make this check in the
future (since it might be somewhat costly). However, the performance was
acceptable during tests, and it had the advantage of changing status
very dynamically. For example, if you change the executable status of
the mcell binary (chmod or simply renaming it to move it while Blender
is running), the status indicator will change by itself whenever the
panel is redrawn (mouse moving over the window for example). Note that
if this dynamic responsiveness isn't needed, then the panel redraw
checking could be replaced with a check when the .blend file is loaded
and also before a run ... just in case.

CellBlender Source ID = a5334e97b66d48faae4e9bb8cbe8ab2096f9a89d