Added Tom's mpl_defaults.py file to set the plotting defaults
Plotting defaults are somewhat of a user preference, but
we should have a standard that we ship with CellBlender.
This version is the one we currently like the best so I
am adding it to Release Candidate 1.0. It *shouldn't* be
a problem since it has been used before, but it wasn't
considered to be part of the released files until now.