Added/updated mouseover descriptions.
Nearly every button, text field, check box, etc now has a mouseover
description. Some existing descriptions were updated for clarity and
consistency.

This update addresses issue 32:
http://code.google.com/p/cellblender/issues/detail?id=32

This is a copy of the text from the issue:

When you mouseover a button, drop-down menu, float property, check box,
or other UI element, a description of what that element does is
presented to the user (assuming the description exists). Many of the UI
elements already have descriptions, but some are still missing. This is
a simple addition that can be very useful to the user. Since there is
very little chance of this breaking anything, I think it should be added
to release_candidate_1.0.