Allow manual viz data loading. Fix abs path bug.
By default, loading viz data works as it did before. However, there is
now an option to manually load viz data as well. This is particularly
useful if you run MCell simulations that weren't created in CellBlender.

If a saved blend file and the related project directory (including viz
data) were moved, this would require the user to re-read the viz data,
upon loading because the path was set absolutely. This was a minor
problem, but it has (mostly) been fixed.

Some comments were added to clarify what happens when viz data is read.