Removed use of the stored parameter dictionary.
For this early version, I decided it was better to regenerate the
parameters from the individual properties rather than try to store a
dictionary for all of the parameters. This simplified the problem and
put the focus back on providing a name/ID translation where it needs to
be for now. The use of a dictionary to store all the parameters and
their values is more of a performance issue which need not be addressed
at this time.

Also, for testing, an earlier version was used to generate files with
large numbers of parameters for testing. Files were created with the
following numbers of parameters:
    10
    20
    50
   100
   500
  1000
  2000
 10000
 50000
Note that the algorithm used to create values was not well-considered
for parameter counts above 1500, so files with more parameters than that
contain many parameters with infinite (inf) values.