Found the recursion ... but not in the recursive algorithm!!
There may be two types of recursion in this code:
  1. Explicit recursion to walk down the parse tree - this works.
  2. Implicit recursion through blender when updating fields
At this time, it appears that the second form of recursion may
have been the problem.

The function "update_parameter_expression" is passed as the
update function to the MCellGeneralParameterProperty.expr property.
But updating a single parameter expression caused this function
to be called over and over again eventually exceeding the recursion
depth.

I don't know what is causing this, but when a global flag was put
into "update_parameter_expression" to limit itself to one call at
a time, the problem disappeared.

I suspect that the problem is being caused because updating one
parameter causes others to change - as it should. But maybe they
are being flagged as "changed" when they've really only been
refilled with the same value. This might be a key to fixing it.

Of course, this entire discussion is speculation, but it adds
up so far.