The previous conjecture appears to be correct.
By limiting the update depth to 10, the recursion error
disappeared. So the problem appears to be that updating
any field causes other fields to be updated even if the
"update" doesn't change anything. If this is true, then
one fix would be to check for actual changes (compare
first) before actually changing a property value.