More work on parameters
Ended up needing a two way dictionary ... currently implemented as
two separate dictionaries.

Everything seems to work except that calling update_parameter_dictionary
when a name is changed isn't the right thing to do (as it was in the
past). So update_parameter_name needs to do something different. Part of
the problem may be that the current code stores the dictionaries, but
doesn't use the stored version and rebuilds them again with every call.
Now that we have a bidirectional mapping between names and IDs, it may
be time to actually use those stored dictionaries as intended.

Note that the use of the term "stored dictionaries" doesn't have to mean
stored in the .blend file. In fact, it might be better if the
dictionaries aren't stored in the .blend file but created from the
properties for each parameter whenever a new .blend file is loaded.

Note that this version of the code is still very rough!!