This version is a first attempt to use locks on the parameter space.
This version "locks up" (no pun intended) when parameters are
modified which indicates that there is an access conflict of some
sort. I might try using reentrant locks (rlocks) to see if that's
the problem or dig into it further. In any case, this version is
NOT suitable for use since it does lock up as soon as you try to
modify a parameter. It's only being committed to keep track of
what has been tried.