Implemented a locking system to keep track of the parameter space.
The locking system "checks out"* a copy of the parameters and uses
that copy to satisfy all requests until all checked out copies have
been checked back in. This version appears to work, but needs lots
of additional testing.

*The notion of "checking out" or "checking in" refers to storing
the parameters (in a "pickled" format) in a Blender Property named:

  mcell.general_parameters.parameter_space_string

This allows a general Python structure to be stored in the .blend
file without the need to code the structure directly in Blender's
limited Properties architecture. In this case, the parameter data
is stored in a "ParameterSpace" object (of class ParameterSpace)
which is not a Blender property.

At this time, it's not clear what the advantages and disadvantages
of this storage approach will be. But it was convenient for this
application because it allowed the "ParameterSpace" class to be
developed (and to actually function) outside of Blender.